backend_api/requirements.txt
# Backend API Dependencies
# FastAPI backend for Smart Retail Analytics

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0

# MQTT
paho-mqtt>=2.0.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# AI Agent (Giai đoạn 6)
google-generativeai>=0.3.0
openai>=1.3.0

